Crate hydro_deploy_integration

Modules

multi_connection

single_connection

Structs

ConnectedDemux

ConnectedDirect

ConnectedTagged

DeployPorts

Contains runtime information passed by Hydro Deploy to a program, describing how to connect to other services and metadata about them.

MergeSource

TaggedSource

Enums

AcceptedServer

BoundServer

ClientConnection

Connection

ServerBindConfig

ServerPort

Describes how to connect to a service which is listening on some port.

Traits

Connected

ConnectedSink

ConnectedSource

StreamSink

Functions

accept_bound

Type Aliases

BufferedDrain

DynSink

DynStream

DynStreamSink

InitConfig